供应链管理项目初始化

+++ bio-drqi-bsm/bio-drqi-bsm-service/src/main/java/com/bio/drqi/bsm/service/impl/BmsStockLocationServiceImpl.java
@@ -10,9 +10,8 @@ import com.bio.drqi.bsm.req.BmsStockLocationEditReqDTO;
 import com.bio.drqi.bsm.req.BmsStockLocationListPageReqDTO;
 import com.bio.drqi.bsm.rsp.BmsStockLocationListAllStockRspDTO;
 import com.bio.drqi.bsm.rsp.BmsStockLocationListPageRspDTO;
-import com.bio.drqi.bsm.rsp.BmsStockLocationQueryByUnitRspDTO;
 import com.bio.drqi.bsm.service.BmsStockLocationService;
-import com.bio.drqi.domain.BmsProductStockOutLog;
+import com.bio.drqi.domain.BmsProductStockTb;
 import com.bio.drqi.domain.BmsStockDict;
 import com.bio.drqi.domain.BmsStockLocationDict;
 import com.bio.drqi.mapper.BmsProductStockTbMapper;
@@ -25,10 +24,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -41,24 +37,9 @@ public class BmsStockLocationServiceImpl implements BmsStockLocationService {
     @Resource
     private BmsStockDictMapper bmsStockDictMapper;
 
-    @Override
-    public List<BmsStockLocationQueryByUnitRspDTO> queryByUnit(String unitCode) {
-        List<BmsStockLocationQueryByUnitRspDTO> bmsStockLocationQueryByUnitRspDTOList = new ArrayList<>();
-        List<BmsStockLocationDict> bmsStockLocationDictList = bmsStockLocationDictMapper.selectAllByUnitCode(unitCode);
-        if (CollectionUtil.isNotEmpty(bmsStockLocationDictList)) {
-            Map<String, List<BmsStockLocationDict>> bmsStockLocationDictMap = bmsStockLocationDictList.stream().collect(Collectors.groupingBy(BmsStockLocationDict::getStockCode));
-            bmsStockLocationDictMap.forEach((stockCode, list) -> {
-                BmsStockLocationQueryByUnitRspDTO bmsStockLocationQueryByUnitRspDTO = new BmsStockLocationQueryByUnitRspDTO();
-                bmsStockLocationQueryByUnitRspDTO.setStockName(list.get(0).getStockName());
-                bmsStockLocationQueryByUnitRspDTO.setStockCode(stockCode);
-                bmsStockLocationQueryByUnitRspDTO.setStockLocationNumber(list.stream().map(BmsStockLocationDict::getLocationNumber).collect(Collectors.toList()));
-                bmsStockLocationQueryByUnitRspDTOList.add(bmsStockLocationQueryByUnitRspDTO);
-            });
+    @Resource
+    private BmsProductStockTbMapper bmsProductStockTbMapper;
 
-
-        }
-        return bmsStockLocationQueryByUnitRspDTOList;
-    }
 
     @Override
     public PageInfo<BmsStockLocationListPageRspDTO> listPage(BmsStockLocationListPageReqDTO bmsStockLocationListPageReqDTO) {
@@ -71,27 +52,18 @@ public class BmsStockLocationServiceImpl implements BmsStockLocationService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void add(BmsStockLocationAddReqDTO bmsStockLocationAddReqDTO) {
-        BmsStockDict bmsStockDict = bmsStockDictMapper.selectOneByStockName(bmsStockLocationAddReqDTO.getStockName().trim());
+        BmsStockDict bmsStockDict = bmsStockDictMapper.selectOneByStockCode(bmsStockLocationAddReqDTO.getStockCode().trim());
         if (bmsStockDict == null) {
-            bmsStockDict = new BmsStockDict();
-            bmsStockDict.setStockName(bmsStockLocationAddReqDTO.getStockName());
-            bmsStockDict.setStockCode(IdUtils.simpleUUID());
-            bmsStockDict.setUnitCode(bmsStockLocationAddReqDTO.getUnitCode());
-            bmsStockDict.setKdNumber(null);
-            bmsStockDict.setCreateTime(new Date());
-            bmsStockDict.setCreateUserId(SecurityContextHolder.getUserId());
-            bmsStockDict.setCreateUserName(SecurityContextHolder.getUserName());
-            bmsStockDictMapper.insert(bmsStockDict);
-        } else {
-            BmsStockLocationDict bmsStockLocationDict = bmsStockLocationDictMapper.selectOneByStockCodeAndLocationNumber(bmsStockDict.getStockCode(), bmsStockLocationAddReqDTO.getLocationNumber());
-            if (bmsStockLocationDict != null) {
-                throw new BusinessException("请不到重复添加库存");
-            }
+            throw new BusinessException("库房不存在");
         }
-        BmsStockLocationDict bmsStockLocationDict = new BmsStockLocationDict();
-        bmsStockLocationDict.setUnitCode(bmsStockLocationAddReqDTO.getUnitCode());
+        BmsStockLocationDict bmsStockLocationDict = bmsStockLocationDictMapper.selectOneByStockCodeAndLocationNumber(bmsStockDict.getStockCode(), bmsStockLocationAddReqDTO.getLocationNumber());
+        if (bmsStockLocationDict != null) {
+            throw new BusinessException("请不到重复添加库存");
+        }
+        bmsStockLocationDict = new BmsStockLocationDict();
+        bmsStockLocationDict.setUnitCode(bmsStockDict.getUnitCode());
         bmsStockLocationDict.setStockCode(bmsStockDict.getStockCode());
-        bmsStockLocationDict.setStockName(bmsStockLocationAddReqDTO.getStockName());
+        bmsStockLocationDict.setStockName(bmsStockDict.getStockName());
         bmsStockLocationDict.setLocationNumber(bmsStockLocationAddReqDTO.getLocationNumber());
         bmsStockLocationDict.setCreateUserId(SecurityContextHolder.getUserId());
         bmsStockLocationDict.setCreateUserName(SecurityContextHolder.getNickName());
@@ -101,6 +73,15 @@ public class BmsStockLocationServiceImpl implements BmsStockLocationService {
 
     @Override
     public void delete(Integer id) {
+        BmsStockLocationDict bmsStockLocationDict = bmsStockLocationDictMapper.selectById(id);
+        if (bmsStockLocationDict == null) {
+            throw new BusinessException("不存在此库位");
+        }
+        List<BmsProductStockTb> bmsProductStockTbList = bmsProductStockTbMapper.selectAllByStockLocationNumberIn(Arrays.asList(bmsStockLocationDict.getLocationNumber()));
+
+        if(CollectionUtil.isNotEmpty(bmsProductStockTbList)){
+            throw new BusinessException("该库位已经使用无法更改");
+        }
         bmsStockLocationDictMapper.deleteById(id);
     }
